[mod] moves list items to their own folder
