+++ src/components/classifier/ClassifierList/ClassifierList.tsx
@@ -9,7 +9,7 @@ import { useDialog, useDialogHotkey } from "hooks";
 import { ImportTensorflowModelDialog } from "components/dialogs";
 import { SaveFittedModelDialog } from "components/dialogs";
 import { CollapsibleList } from "components/common/styled-components/CollapsibleList";
-import { ClassifierExecListItem } from "../ClassifierExecListItem";
+import { ClassifierExecListItem } from "components/list-items";
 import { CategoriesList } from "components/categories/CategoriesList";
 
 import {
