+++ src/components/common/styled-components/ApplicationOptionsList.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { List } from "@mui/material";
 
-import { SettingsListItem } from "../list-items/SettingsListItem";
+import { SettingsListItem } from "../../list-items/SettingsListItem";
 import { HelpDrawer } from "../../help";
-import { SendFeedbackListItem } from "../list-items/SendFeedbackListItem";
+import { SendFeedbackListItem } from "../../list-items/SendFeedbackListItem";
 
 export const ApplicationOptionsList = () => {
   return (
