--- src/components/file-io/list-items/OpenListItem/index.ts
@@ -1,1 +1,0 @@
-export { OpenListItem } from "./OpenListItem";
