--- src/components/file-io/list-items/SaveListItem/index.ts
@@ -1,1 +1,0 @@
-export { SaveListItem } from "./SaveListItem";
