+++ src/components/list-items/NewProjectListItem/NewProjectListItem.tsx
