+++ src/components/list-items/OpenListItem/OpenListItem.tsx
@@ -1,10 +1,7 @@
-import React from "react";
-
 import { ListItem, ListItemIcon, ListItemText } from "@mui/material";
-
 import FolderOpenIcon from "@mui/icons-material/FolderOpen";
 
-import { OpenMenu } from "../../menus/OpenMenu";
+import { OpenMenu } from "../../file-io/menus/OpenMenu";
 import { useMenu } from "hooks";
 
 export const OpenListItem = () => {
