+++ src/components/segmenter/ArchitectureSettingsGrid/SegmenterArchitectureSettingsGrid.tsx
